requirements.txt
streamlit
pandas
requests
gdown
gspread
google-auth
python-dotenv
pydub